99% implemented now!
Had to change the way the internal buffers are extracted, and how the
algorithm responds when it could not find enough unique values to fill
the buffers.

These changes make TestingRandomFew run 75% as fast as stable_sort now,
while before it was only 15% (!) as fast. Still need to clean up the
code, possibly optimize here and there, and try out different
algorithms for the third branch in Merge to see which is the best.