Stable sorting network = super fast!
I switched the binary insertion sort over to an unstable sorting
network (turned into a stable one using a fixed-size external buffer)
to see if it was any faster, and... yes, it's a lot faster! It's
actually ALMOST as fast as std::stable_sort() in all cases now, and a
lot faster than my merge sort implementation!