Optimizing for small arrays
WikiSort was a lot slower than std::stable_sort() for 0-32 items due to
its use of InsertionSort. It now uses sorting networks for arrays of up
to size 8, and for larger arrays the in-place merge algorithm had to be
tweaked to support tiny subarray sizes. It's a bit more complex, but
small arrays sort much faster now without affecting the sort speed for
larger arrays.