Technically this specific implementation is faster
The read me said block sort is generally as fast as a merge sort, but that's actually not true. It derives most of its speed from using a ton of optimizations at the lower levels of the sort, and from using a decent-sized cache stored on the stack. Block merging on its own – without any optimizations – is generally about 60%-65% as fast as a merge sort.