Slightly faster than the old version
For some reason using std::rotate() on its own is faster than using the cache for small rotations? Using the cache is a lot faster in the C version. I tried copying and pasting the source code for std::rotate() into WikiSort but it still wasn't faster than the cached version. Not sure what causes that.

Anyway, this also tags the first value of each A block and uses modified versions of MergeInPlace and MergeInto. The main speed improvements were from using std::rotate(), though.